Fix dev:check fail-fast not returning error (#11672)
Fix dantherm definition

+++ tests/bootstrap.php
@@ -68,10 +68,12 @@ if (getenv('DBTEST')) {
 
     // try to avoid erasing people's primary databases
     if ($db_config['database'] !== \config('database.connections.mysql.database', 'librenms')) {
-        echo "Refreshing database...";
-        $migrate_result = Artisan::call('migrate:fresh', ['--seed' => true, '--env' => 'testing', '--database' => 'testing']);
-        $migrate_output = Artisan::output();
-        echo "done\n";
+        if (!getenv('SKIP_DB_REFRESH')) {
+            echo "Refreshing database...";
+            $migrate_result = Artisan::call('migrate:fresh', ['--seed' => true, '--env' => 'testing', '--database' => 'testing']);
+            $migrate_output = Artisan::output();
+            echo "done\n";
+        }
     } else {
         echo "Info: Refusing to reset main database: {$db_config['database']}.  Running migrations.\n";
         $migrate_result = Artisan::call('migrate', ['--seed' => true, '--env' => 'testing', '--database' => 'testing']);
